refactor(ui): format UI components, view models, and navigation

+++ risc_tool/ui/data_explorer/outlier.py
@@ -45,7 +45,8 @@ def show_boxplot(variable: str):
     )
 
     boxplot = (
-        alt.Chart(df)
+        alt
+        .Chart(df)
         .mark_boxplot(
             orient="horizontal",
             size=20,
@@ -56,7 +57,8 @@ def show_boxplot(variable: str):
     )
 
     anchors = (
-        alt.Chart(perc_df)
+        alt
+        .Chart(perc_df)
         .mark_point(size=300, opacity=0)
         .encode(
             x="Value:Q", tooltip=["Percentile:N", alt.Tooltip("Value:Q", format=".2f")]
@@ -65,7 +67,8 @@ def show_boxplot(variable: str):
     )
 
     arrows = (
-        alt.Chart(perc_df)
+        alt
+        .Chart(perc_df)
         .mark_text(text="▼", size=16, baseline="bottom", dy=0)
         .encode(
             x="Value:Q",
@@ -74,7 +77,8 @@ def show_boxplot(variable: str):
     )
 
     text_above = (
-        alt.Chart(perc_df)
+        alt
+        .Chart(perc_df)
         .mark_text(
             align="center",
             baseline="bottom",
@@ -90,7 +94,8 @@ def show_boxplot(variable: str):
     )
 
     text_below = (
-        alt.Chart(perc_df)
+        alt
+        .Chart(perc_df)
         .mark_text(
             align="center",
             baseline="top",
@@ -268,17 +273,16 @@ def outlier_rule_input(
                     )
                     st.rerun()
 
-                if outlier_id != FilterID.TEMPORARY:
-                    if st.button(
-                        label="Delete Rule",
-                        type="primary",
-                        key=f"outlier-delete-button-{key}",
-                        icon=":material/delete:",
-                        width="stretch",
-                    ):
-                        logger.info("User deleted outlier rule ID %s", outlier_id)
-                        de_view_model.delete_outlier_rule(outlier_id)
-                        st.rerun()
+                if outlier_id != FilterID.TEMPORARY and st.button(
+                    label="Delete Rule",
+                    type="primary",
+                    key=f"outlier-delete-button-{key}",
+                    icon=":material/delete:",
+                    width="stretch",
+                ):
+                    logger.info("User deleted outlier rule ID %s", outlier_id)
+                    de_view_model.delete_outlier_rule(outlier_id)
+                    st.rerun()
 
 
 def outlier_rules():
